LeetCode Daily Challenge: 2021-03-06 Solution
Pretty terrible brute force solution. I was able to grasp that presorting the words by length in reverse order helped with runtime, but I completely missed that my inner check could have been improved greatly if I started to preform the encoded string as it would be in a simple in check rather that the current 2nd loop. Other smarter solutions seemed to be using a trie data structure which was also completely missed by me.