LeetCode 478. Generate Random Point in a Circle
Solved for Daily Challenge (2021-03-17).

This was an annoying problem. I was able to recollect enough from geometry to solve for the most of it and it was passing small test cases. The problem was when they scaled up the number of random points it was no longer considered uniformly distributed. This basically required an addition of a sqrt to make sure higher distances from the center were more favored. Probably couldn't have gotten there without external info though about this topic. Good to know for the future.